wip: add new folder btn and functionality
