+++ src/components/ui/FileBrowser/FileList.tsx
@@ -1,17 +1,9 @@
 import * as React from 'react';
-import { IconButton, Typography } from '@material-tailwind/react';
-import {
-  DocumentIcon,
-  EllipsisHorizontalCircleIcon,
-  FolderIcon
-} from '@heroicons/react/24/outline';
+import { Typography } from '@material-tailwind/react';
 
 import type { File, FileSharePathItem } from '../../../shared.types';
 import FileListCrumbs from './Crumbs';
-import { useFileBrowserContext } from '../../../contexts/FileBrowserContext';
-import { useZoneBrowserContext } from '../../../contexts/ZoneBrowserContext';
-import useHandleLeftClick from '../../../hooks/useHandleLeftClick';
-import { formatDate, formatFileSize } from '../../../utils';
+import FileRow from './FileRow';
 
 type FileListProps = {
   displayFiles: File[];
@@ -47,9 +39,6 @@ export default function FileList({
   setPropertiesTarget,
   handleRightClick
 }: FileListProps): JSX.Element {
-  const { fetchAndFormatFilesForDisplay } = useFileBrowserContext();
-  const { handleLeftClick } = useHandleLeftClick();
-  const { currentFileSharePath } = useZoneBrowserContext();
   return (
     <div
       className={`px-2 transition-all duration-300 ${showPropertiesDrawer ? 'mr-[350px]' : ''}`}
@@ -84,111 +73,18 @@ export default function FileList({
         {/* File rows */}
         {displayFiles.length > 0 &&
           displayFiles.map((file, index) => {
-            const isSelected = selectedFiles.some(
-              selectedFile => selectedFile.name === file.name
-            );
-
             return (
-              <div
+              <FileRow
                 key={file.name}
-                className={`cursor-pointer min-w-fit grid grid-cols-[minmax(170px,2fr)_minmax(80px,1fr)_minmax(95px,1fr)_minmax(75px,1fr)_minmax(40px,1fr)] gap-4 hover:bg-primary-light/30 focus:bg-primary-light/30 ${isSelected && 'bg-primary-light/30'} ${index % 2 === 0 && !isSelected && 'bg-surface/50'}  `}
-                onClick={(e: React.MouseEvent<HTMLDivElement>) =>
-                  handleLeftClick(
-                    e,
-                    file,
-                    selectedFiles,
-                    setSelectedFiles,
-                    displayFiles,
-                    currentFileSharePath,
-                    setPropertiesTarget,
-                    showPropertiesDrawer
-                  )
-                }
-                onContextMenu={(e: React.MouseEvent<HTMLDivElement>) =>
-                  handleRightClick(
-                    e,
-                    file,
-                    selectedFiles,
-                    setSelectedFiles,
-                    currentFileSharePath,
-                    setPropertiesTarget
-                  )
-                }
-                onDoubleClick={() => {
-                  if (file.is_dir && currentFileSharePath) {
-                    fetchAndFormatFilesForDisplay(
-                      `${currentFileSharePath.name}?subpath=${file.path}`
-                    );
-                  }
-                }}
-              >
-                {/* Name column */}
-                <div className="flex items-center w-full gap-3 pl-3 py-1 overflow-x-auto">
-                  <Typography
-                    variant="small"
-                    className="font-medium text-primary-light hover:underline"
-                    onClick={(e: React.MouseEvent<HTMLDivElement>) => {
-                      e.stopPropagation();
-                      if (file.is_dir) {
-                        fetchAndFormatFilesForDisplay(
-                          `${currentFileSharePath?.name}?subpath=${file.path}`
-                        );
-                        setPropertiesTarget({
-                          targetFile: file,
-                          fileSharePath: currentFileSharePath
-                        });
-                      }
-                    }}
-                  >
-                    {file.name}
-                  </Typography>
-                </div>
-
-                {/* Type column */}
-                <div className="flex items-center w-full gap-3 py-1 text-grey-700 overflow-x-auto">
-                  {file.is_dir ? (
-                    <FolderIcon className="text-foreground h-5 w-5" />
-                  ) : (
-                    <DocumentIcon className="text-foreground h-5 w-5" />
-                  )}
-                  <Typography variant="small" className="font-medium">
-                    {file.is_dir ? 'Folder' : 'File'}
-                  </Typography>
-                </div>
-
-                {/* Last Modified column */}
-                <div className="py-1 text-grey-700  flex items-center overflow-x-auto">
-                  <Typography variant="small" className="font-medium">
-                    {formatDate(file.last_modified)}
-                  </Typography>
-                </div>
-
-                {/* Size column */}
-                <div className="py-1 text-grey-700 flex items-center overflow-x-auto">
-                  <Typography variant="small" className="font-medium">
-                    {file.is_dir ? '—' : formatFileSize(file.size)}
-                  </Typography>
-                </div>
-
-                {/* Context menu button */}
-                <div
-                  className="py-1 text-grey-700 flex items-center flex-shrink-0"
-                  onClick={e => {
-                    handleRightClick(
-                      e,
-                      file,
-                      selectedFiles,
-                      setSelectedFiles,
-                      currentFileSharePath,
-                      setPropertiesTarget
-                    );
-                  }}
-                >
-                  <IconButton variant="ghost">
-                    <EllipsisHorizontalCircleIcon className="h-5 w-5 text-foreground" />
-                  </IconButton>
-                </div>
-              </div>
+                file={file}
+                index={index}
+                selectedFiles={selectedFiles}
+                setSelectedFiles={setSelectedFiles}
+                displayFiles={displayFiles}
+                showPropertiesDrawer={showPropertiesDrawer}
+                setPropertiesTarget={setPropertiesTarget}
+                handleRightClick={handleRightClick}
+              />
             );
           })}
       </div>
